LPS-75002 Move RestrictedByteArrayCacheOutputStream to com.liferay.portal.kernel.servlet.internal.
